Move thlo tiling tests out of gml_st and use upstream interface.
PiperOrigin-RevId: 510968582

+++ tensorflow/compiler/xla/mlir_hlo/gml_st/transforms/scalarization/scalarization.cc
@@ -168,7 +168,7 @@ struct ScalarizeScatterOp : public OpRewritePattern<thlo::ScatterOp> {
     Value updates = scatterOp.getUpdates();
     auto updatesType = updates.getType().dyn_cast<RankedTensorType>();
     if (!updatesType) return failure();
-    int64_t updatesRank = updatesType.getRank();
+    unsigned updatesRank = updatesType.getRank();
 
     SmallVector<OpFoldResult> updatesDimSizes =
         tensor::getMixedSizes(b, loc, updates);
